# feat: function generator