--- a/src/index.ts
+++ b/src/index.ts
@@ -1,3 +1,3 @@
 export { default as transform } from './transformers'
-export { default as matches } from './matchers'
+export { default as matches, getMatchFunction } from './matchers'
 export { default as Store } from './store'