… coverage/clover.xml
… coverage/coverage-final.json
… coverage/lcov-report/favicon.png
… coverage/lcov-report/index.html
… coverage/lcov-report/index.ts.html
… coverage/lcov-report/matchers.ts.html
… coverage/lcov-report/prettify.css
… coverage/lcov-report/prettify.js
… coverage/lcov-report/sort-arrow-sprite.png
… coverage/lcov-report/store.ts.html
… coverage/lcov-report/transformers.ts.html
… coverage/lcov-report/unset.ts.html
… coverage/lcov.info
… src/__tests__/matchers.test.ts
… src/store.ts
… src/test.json